refactor: ToolbarActions
