+++ src/components/dashboard/types.tsx
@@ -1,0 +1,11 @@
+// Common types used in dashboard
+export type Props = {
+    settingsLabel: string,
+    modeTitle: string,
+    lightLabel: string,
+    darkLabel: string,
+    systemLabel: string,
+    languageTitle: string,
+    frenchLabel: string,
+    englishLabel: string
+}
